requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
gemini-ai
spacy
PyPDF2
docx2txt
sentence-transformers
faiss-cpu
psycopg2-binary